refactor(compliance): refactor and document compliance modules
